# exchange-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exchange Rate Application",
    "description": "This is an exchange rate application where users can sign up for an account, add transactions, see their previous transactions, and get the current exchange rate calculated according to the past 3 days. Users can also check the conversion rate and amount for buying or selling USD and LBP.",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/"
    }
  ],
...
  "paths": {
...
    "/statistics/getExtremums": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get extremums of exchange rates within a specific date range",
        "operationId": "getExtremums",
        "parameters": [
          {
            "name": "START_YEAR",
            "in": "query",
            "required": true,
            "description": "Start year, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "START_MONTH",
            "in": "query",
            "required": true,
            "description": "Start month, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "START_DAY",
            "in": "query",
            "required": true,
            "description": "Start day, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 12
            }
          },
          {
            "name": "END_YEAR",
            "in": "query",
            "required": true,
            "description": "End year, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "END_MONTH",
            "in": "query",
            "required": true,
            "description": "End month, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "END_DAY",
            "in": "query",
            "required": true,
            "description": "End day, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "min_lbp_to_usd": {
                      "type": "number",
                      "format": "number",
                      "example": 85000.12
                    },
                    "max_lbp_to_usd": {
                      "type": "number",
                      "format": "number",
                      "example": 90000.34
                    },
                    "min_lbp_to_usd_date": {
                      "type": "string",
                      "format": "date",
                      "example": "01-04-2024"
                    },
                    "max_lbp_to_usd_date": {
                      "type": "string",
                      "format": "date",
                      "example": "03-04-2024"
                    },
                    "min_usd_to_lbp": {
                      "type": "number",
                      "format": "number",
                      "example": 87000
                    },
                    "max_usd_to_lbp": {
                      "type": "number",
                      "format": "number",
                      "example": 91000
                    },
                    "min_usd_to_lbp_date": {
                      "type": "string",
                      "format": "date",
                      "example": "05-04-2024"
                    },
                    "max_usd_to_lbp_date": {
                      "type": "string",
                      "format": "date",
                      "example": "10-04-2024"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters provided"
          }
        }
      }
    },
    "/statistics/getAverageRates": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get average exchange rates within a specific date range",
        "operationId": "getAverageRates",
        "parameters": [
          {
            "name": "START_YEAR",
            "in": "query",
            "required": true,
            "description": "Start year, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "START_MONTH",
            "in": "query",
            "required": true,
            "description": "Start month, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "START_DAY",
            "in": "query",
            "required": true,
            "description": "Start day, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 12
            }
          },
          {
            "name": "END_YEAR",
            "in": "query",
            "required": true,
            "description": "End year, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "END_MONTH",
            "in": "query",
            "required": true,
            "description": "End month, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "END_DAY",
            "in": "query",
            "required": true,
            "description": "End day, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "avg_lbp_to_usd": {
                      "type": "number",
                      "format": "number",
                      "example": 87500.23
                    },
                    "avg_usd_to_lbp": {
                      "type": "number",
                      "format": "number",
                      "example": 89000
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters provided"
          }
        }
      }
    },
    "/statistics/getPercentChange": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get percentage change in exchange rates within a specific date range",
        "operationId": "getPercentChange",
        "parameters": [
          {
            "name": "START_YEAR",
            "in": "query",
            "required": true,
            "description": "Start year, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "START_MONTH",
            "in": "query",
            "required": true,
            "description": "Start month, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "START_DAY",
            "in": "query",
            "required": true,
            "description": "Start day, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 12
            }
          },
          {
            "name": "END_YEAR",
            "in": "query",
            "required": true,
            "description": "End year, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "END_MONTH",
            "in": "query",
            "required": true,
            "description": "End month, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "END_DAY",
            "in": "query",
            "required": true,
            "description": "End day, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "percent_lbp_to_usd": {
                      "type": "number",
                      "format": "number",
                      "example": 5.2
                    },
                    "percent_usd_to_lbp": {
                      "type": "number",
                      "format": "number",
                      "example": -3.4
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters provided"
          }
        }
      }
    },
    "/statistics/getVolumes": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get volumes of exchange transactions within a specific date range",
        "operationId": "getVolumes",
        "parameters": [
          {
            "name": "START_YEAR",
            "in": "query",
            "required": true,
            "description": "Start year, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "START_MONTH",
            "in": "query",
            "required": true,
            "description": "Start month, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "START_DAY",
            "in": "query",
            "required": true,
            "description": "Start day, if null, first date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 12
            }
          },
          {
            "name": "END_YEAR",
            "in": "query",
            "required": true,
            "description": "End year, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "END_MONTH",
            "in": "query",
            "required": true,
            "description": "End month, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "END_DAY",
            "in": "query",
            "required": true,
            "description": "End day, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lbp_to_usd_volume": {
                      "type": "integer",
                      "example": 10
                    },
                    "usd_to_lbp_volume": {
                      "type": "integer",
                      "example": 15
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters provided"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "JWTToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "JSON Web Token (JWT) allow users to link transactions to themselves.\nAdd the JWT token in the format `Bearer <token>`.\n"
      },
      "None": {
        "type": "http",
        "scheme": "none",
        "description": "Authentication is not needed always.\n"
      }
    }
  }
}